ProFTPD: remove <Anonymous> entirely
In fact, the whole <Anonymous>...</Anonymous> section of the config file
is not needed to access and trigger the backdoor. The least amount of
configuration should be used here: the user can still log in and perform
all the usual FTP tasks if they know the right USER/PASS combo for the
machine, but otherwise they are restricted.