Belkin: replace system() with fork + exec
In order to have the command run by the backdoor not inherit the
parent's environment, the system() call is replaced with a fork and an
execle. This has the exact same behavior, except that the child process
does not inherit any environment variables set in the parent.

This modification is tentatively made to get around a hard-to-resolve
issue with preloads when fuzzing this target.